class refactoring
