+++ src/test/java/org/broadinstitute/consent/http/models/DatasetRegistrationSchemaV1BuilderTest.java
@@ -1,5 +1,36 @@
 package org.broadinstitute.consent.http.models;
 
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.alternativeDataSharingPlanControlledOpenAccess;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.alternativeDataSharingPlanDataReleased;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.alternativeDataSharingPlanDataSubmitted;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.alternativeDataSharingPlanExplanation;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.alternativeDataSharingPlanFileName;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.alternativeDataSharingPlanReasons;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.alternativeDataSharingPlanTargetDeliveryDate;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.alternativeDataSharingPlanTargetPublicReleaseDate;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.collaboratingSites;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.controlledAccessRequiredForGenomicSummaryResultsGSR;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.controlledAccessRequiredForGenomicSummaryResultsGSRRequiredExplanation;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.dataCustodianEmail;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.dbGaPPhsID;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.dbGaPStudyRegistrationName;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.embargoReleaseDate;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.generalResearchUse;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.hmb;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.multiCenterStudy;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.nihAnvilUse;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.nihGenomicProgramAdministratorName;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.nihGrantContractNumber;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.nihICsSupportingStudy;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.nihInstitutionCenterSubmission;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.nihProgramOfficerName;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.openAccess;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.phenotypeIndication;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.piInstitution;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.sequencingCenter;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.species;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.studyType;
+import static org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder.submittingToAnvil;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
@@ -19,8 +50,8 @@ import org.broadinstitute.consent.http.models.dataset_registration_v1.DatasetReg
 import org.broadinstitute.consent.http.models.dataset_registration_v1.DatasetRegistrationSchemaV1.NihAnvilUse;
 import org.broadinstitute.consent.http.models.dataset_registration_v1.DatasetRegistrationSchemaV1.NihInstitutionCenterSubmission;
 import org.broadinstitute.consent.http.models.dataset_registration_v1.DatasetRegistrationSchemaV1.StudyType;
-import org.broadinstitute.consent.http.models.dataset_registration_v1.DatasetRegistrationSchemaV1Builder;
 import org.broadinstitute.consent.http.models.dataset_registration_v1.NihICsSupportingStudy;
+import org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder;
 import org.broadinstitute.consent.http.util.gson.GsonUtil;
 import org.junit.jupiter.api.Test;
 
@@ -77,10 +108,6 @@ class DatasetRegistrationSchemaV1BuilderTest {
     assertNotNull(schemaV1.getAlternativeDataSharingPlanTargetDeliveryDate());
     assertNotNull(schemaV1.getAlternativeDataSharingPlanTargetPublicReleaseDate());
     assertNotNull(schemaV1.getAlternativeDataSharingPlanControlledOpenAccess());
-
-    // TODO: Test code ... remove when complete.
-    Gson gson = new GsonBuilder().setPrettyPrinting().create();
-    System.out.println(gson.toJson(schemaV1));
   }
 
   @Test
@@ -122,34 +149,34 @@ class DatasetRegistrationSchemaV1BuilderTest {
   }
 
   private void addAllStudyProperties(Study study) {
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.studyType, study.getStudyId(), StudyType.OBSERVATIONAL.value(), PropertyType.String));
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.phenotypeIndication, study.getStudyId(), randomString(), PropertyType.String));
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.species, study.getStudyId(), randomString(), PropertyType.String));
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.dataCustodianEmail, study.getStudyId(), GsonUtil.getInstance().toJson(List.of(randomString())), PropertyType.Json));
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.nihAnvilUse, study.getStudyId(), NihAnvilUse.I_AM_NHGRI_FUNDED_AND_I_HAVE_A_DB_GA_P_PHS_ID_ALREADY.value(), PropertyType.String));
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.submittingToAnvil, study.getStudyId(), Boolean.TRUE, PropertyType.Boolean));
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.dbGaPPhsID, study.getStudyId(), randomString(), PropertyType.String));
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.dbGaPStudyRegistrationName, study.getStudyId(), randomString(), PropertyType.String));
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.embargoReleaseDate, study.getStudyId(), randomString(), PropertyType.String));
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.sequencingCenter, study.getStudyId(), randomString(), PropertyType.String));
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.piInstitution, study.getStudyId(), randomInt(), PropertyType.Number));
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.nihGrantContractNumber, study.getStudyId(), randomString(), PropertyType.String));
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.nihICsSupportingStudy, study.getStudyId(), GsonUtil.getInstance().toJson(List.of(NihICsSupportingStudy.CC.value())), PropertyType.Json));
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.nihProgramOfficerName, study.getStudyId(), randomString(), PropertyType.String));
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.nihInstitutionCenterSubmission, study.getStudyId(), NihInstitutionCenterSubmission.CC.value(), PropertyType.String));
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.nihGenomicProgramAdministratorName, study.getStudyId(), randomString(), PropertyType.String));
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.multiCenterStudy, study.getStudyId(), Boolean.TRUE, PropertyType.Boolean));
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.collaboratingSites, study.getStudyId(), GsonUtil.getInstance().toJson(List.of(randomString())), PropertyType.Json));
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.controlledAccessRequiredForGenomicSummaryResultsGSR, study.getStudyId(), Boolean.TRUE, PropertyType.Boolean));
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.controlledAccessRequiredForGenomicSummaryResultsGSRRequiredExplanation, study.getStudyId(), randomString(), PropertyType.String));
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.alternativeDataSharingPlanReasons, study.getStudyId(), GsonUtil.getInstance().toJson(List.of(AlternativeDataSharingPlanReason.OTHER.value())), PropertyType.Json));
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.alternativeDataSharingPlanExplanation, study.getStudyId(), randomString(), PropertyType.String));
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.alternativeDataSharingPlanFileName, study.getStudyId(), randomString(), PropertyType.String));
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.alternativeDataSharingPlanDataSubmitted, study.getStudyId(), AlternativeDataSharingPlanDataSubmitted.BY_BATCHES_OVER_STUDY_TIMELINE_E_G_BASED_ON_CLINICAL_TRIAL_ENROLLMENT_BENCHMARKS.value(), PropertyType.String));
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.alternativeDataSharingPlanDataReleased, study.getStudyId(), Boolean.TRUE, PropertyType.Boolean));
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.alternativeDataSharingPlanTargetDeliveryDate, study.getStudyId(), randomString(), PropertyType.String));
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.alternativeDataSharingPlanTargetPublicReleaseDate, study.getStudyId(), randomString(), PropertyType.String));
-    study.addProperty(createStudyProperty(DatasetRegistrationSchemaV1Builder.alternativeDataSharingPlanControlledOpenAccess, study.getStudyId(), AlternativeDataSharingPlanControlledOpenAccess.OPEN_ACCESS.value(), PropertyType.String));
+    study.addProperty(createStudyProperty(studyType, study.getStudyId(), StudyType.OBSERVATIONAL.value(), PropertyType.String));
+    study.addProperty(createStudyProperty(phenotypeIndication, study.getStudyId(), randomString(), PropertyType.String));
+    study.addProperty(createStudyProperty(species, study.getStudyId(), randomString(), PropertyType.String));
+    study.addProperty(createStudyProperty(dataCustodianEmail, study.getStudyId(), GsonUtil.getInstance().toJson(List.of(randomString())), PropertyType.Json));
+    study.addProperty(createStudyProperty(nihAnvilUse, study.getStudyId(), NihAnvilUse.I_AM_NHGRI_FUNDED_AND_I_HAVE_A_DB_GA_P_PHS_ID_ALREADY.value(), PropertyType.String));
+    study.addProperty(createStudyProperty(submittingToAnvil, study.getStudyId(), Boolean.TRUE, PropertyType.Boolean));
+    study.addProperty(createStudyProperty(dbGaPPhsID, study.getStudyId(), randomString(), PropertyType.String));
+    study.addProperty(createStudyProperty(dbGaPStudyRegistrationName, study.getStudyId(), randomString(), PropertyType.String));
+    study.addProperty(createStudyProperty(embargoReleaseDate, study.getStudyId(), randomString(), PropertyType.String));
+    study.addProperty(createStudyProperty(sequencingCenter, study.getStudyId(), randomString(), PropertyType.String));
+    study.addProperty(createStudyProperty(piInstitution, study.getStudyId(), randomInt(), PropertyType.Number));
+    study.addProperty(createStudyProperty(nihGrantContractNumber, study.getStudyId(), randomString(), PropertyType.String));
+    study.addProperty(createStudyProperty(nihICsSupportingStudy, study.getStudyId(), GsonUtil.getInstance().toJson(List.of(NihICsSupportingStudy.CC.value())), PropertyType.Json));
+    study.addProperty(createStudyProperty(nihProgramOfficerName, study.getStudyId(), randomString(), PropertyType.String));
+    study.addProperty(createStudyProperty(nihInstitutionCenterSubmission, study.getStudyId(), NihInstitutionCenterSubmission.CC.value(), PropertyType.String));
+    study.addProperty(createStudyProperty(nihGenomicProgramAdministratorName, study.getStudyId(), randomString(), PropertyType.String));
+    study.addProperty(createStudyProperty(multiCenterStudy, study.getStudyId(), Boolean.TRUE, PropertyType.Boolean));
+    study.addProperty(createStudyProperty(collaboratingSites, study.getStudyId(), GsonUtil.getInstance().toJson(List.of(randomString())), PropertyType.Json));
+    study.addProperty(createStudyProperty(controlledAccessRequiredForGenomicSummaryResultsGSR, study.getStudyId(), Boolean.TRUE, PropertyType.Boolean));
+    study.addProperty(createStudyProperty(controlledAccessRequiredForGenomicSummaryResultsGSRRequiredExplanation, study.getStudyId(), randomString(), PropertyType.String));
+    study.addProperty(createStudyProperty(alternativeDataSharingPlanReasons, study.getStudyId(), GsonUtil.getInstance().toJson(List.of(AlternativeDataSharingPlanReason.OTHER.value())), PropertyType.Json));
+    study.addProperty(createStudyProperty(alternativeDataSharingPlanExplanation, study.getStudyId(), randomString(), PropertyType.String));
+    study.addProperty(createStudyProperty(alternativeDataSharingPlanFileName, study.getStudyId(), randomString(), PropertyType.String));
+    study.addProperty(createStudyProperty(alternativeDataSharingPlanDataSubmitted, study.getStudyId(), AlternativeDataSharingPlanDataSubmitted.BY_BATCHES_OVER_STUDY_TIMELINE_E_G_BASED_ON_CLINICAL_TRIAL_ENROLLMENT_BENCHMARKS.value(), PropertyType.String));
+    study.addProperty(createStudyProperty(alternativeDataSharingPlanDataReleased, study.getStudyId(), Boolean.TRUE, PropertyType.Boolean));
+    study.addProperty(createStudyProperty(alternativeDataSharingPlanTargetDeliveryDate, study.getStudyId(), randomString(), PropertyType.String));
+    study.addProperty(createStudyProperty(alternativeDataSharingPlanTargetPublicReleaseDate, study.getStudyId(), randomString(), PropertyType.String));
+    study.addProperty(createStudyProperty(alternativeDataSharingPlanControlledOpenAccess, study.getStudyId(), AlternativeDataSharingPlanControlledOpenAccess.OPEN_ACCESS.value(), PropertyType.String));
   }
 
   private String randomString() {
@@ -191,9 +218,10 @@ class DatasetRegistrationSchemaV1BuilderTest {
   }
 
   private void addAllDatasetProperties(Dataset dataset) {
-    dataset.addProperty(createDatasetProperty(dataset, DatasetRegistrationSchemaV1Builder.openAccess, PropertyType.Boolean));
-    dataset.addProperty(createDatasetProperty(dataset, DatasetRegistrationSchemaV1Builder.generalResearchUse, PropertyType.Boolean));
-    dataset.addProperty(createDatasetProperty(dataset, DatasetRegistrationSchemaV1Builder.hmb, PropertyType.Boolean));
+    dataset.addProperty(createDatasetProperty(dataset, openAccess, PropertyType.Boolean));
+    dataset.addProperty(createDatasetProperty(dataset, generalResearchUse, PropertyType.Boolean));
+    dataset.addProperty(createDatasetProperty(dataset, hmb, PropertyType.Boolean));
+    // TODO: Flesh out all possible properties
   }
 
   private DatasetProperty createDatasetProperty(Dataset dataset, String schemaProp, PropertyType type) {
